Add an example demonstrating use of the process APIs.
Also, move `getcwd` into the `process` module, since it's reading an
attribute of the current process. This is admittadly subjective, since
`cwd()` is in `fs`, but I think it makes more sense this way because
`cwd()` is just a way to let `*at` functions behave like their non-`at`
counterparts, while `getcwd` is typically used to record the path for
later use or for use outside the process.
