+++ src/fs/cwd.rs
@@ -1,7 +1,5 @@
-use std::ffi::OsString;
-
 use crate::imp;
-use crate::io::{self, RawFd};
+use crate::io::RawFd;
 use io_lifetimes::BorrowedFd;
 
 /// `AT_FDCWD`—Returns a handle representing the current working directory.
@@ -27,39 +25,5 @@ pub fn cwd() -> BorrowedFd<'static> {
     #[allow(unsafe_code)]
     unsafe {
         BorrowedFd::<'static>::borrow_raw_fd(at_fdcwd)
-    }
-}
-
-/// `getcwd()`
-///
-/// If `reuse` is non-empty, reuse its buffer to store the result if possible.
-///
-/// # References
-///  - [POSIX]
-///  - [Linux]
-///
-/// [POSIX]: https://pubs.opengroup.org/onlinepubs/9699919799/functions/getcwd.html
-/// [Linux]: https://man7.org/linux/man-pages/man3/getcwd.3.html
-#[cfg(not(target_os = "wasi"))]
-#[inline]
-pub fn getcwd(reuse: OsString) -> io::Result<OsString> {
-    use std::os::unix::prelude::OsStringExt;
-
-    // This code would benefit from having a better way to read into
-    // uninitialized memory, but that requires `unsafe`.
-    let mut buffer = reuse.into_vec();
-    buffer.clear();
-    buffer.resize(256, 0_u8);
-
-    loop {
-        match imp::syscalls::getcwd(&mut buffer) {
-            Err(imp::io::Error::RANGE) => buffer.resize(buffer.len() * 2, 0_u8),
-            Ok(_) => {
-                let len = buffer.iter().position(|x| *x == b'\0').unwrap();
-                buffer.resize(len, 0_u8);
-                return Ok(OsString::from_vec(buffer));
-            }
-            Err(errno) => return Err(errno),
-        }
     }
 }
