+++ src/fs/mod.rs
@@ -81,8 +81,6 @@ pub use constants::{Access, FdFlags, Mode, OFlags};
 pub use copy_file_range::copy_file_range;
 #[cfg(not(target_os = "redox"))]
 pub use cwd::cwd;
-#[cfg(not(any(target_os = "redox", target_os = "wasi")))]
-pub use cwd::getcwd;
 #[cfg(not(target_os = "redox"))]
 pub use dir::{Dir, DirEntry};
 #[cfg(not(any(
